[refactor]#108 유저 게시글 interface 분리
